This implementation saves the offset from the start of the cycle instead
of from the previous step, called jumps later. This favores big numerator segments as can
be seen by the benchmarks.

I decided to drop the offsets and just use the jumps
to get the offsets as the most CPU intensive part of this is calculating the jumps
Obviously this means that if most uses cases require the offsets this is not better performing

I would expect that in the cases that this gets worsen it gets worsen insignificantly
compared to the cases where the new "jumps" give much better perfomance

The sort.Search is inlined because this gives something like 30-70%
boost which I would argue here it is worth the few extra lines

name                                                                                                                  old time/op    new time/op    delta
GetStripedOffsets/length10,seed777-8                                                                                    34.5µs ±23%    25.8µs ±12%   -25.27%  (p=0.008 n=5+5)
GetStripedOffsets/length100,seed777-8                                                                                   1.65ms ±15%    1.65ms ± 8%      ~     (p=0.690 n=5+5)
GetStripedOffsetsEven/length10-8                                                                                        3.85µs ± 2%    3.75µs ± 4%      ~     (p=0.151 n=5+5)
GetStripedOffsetsEven/length100-8                                                                                       53.5µs ± 1%    50.9µs ± 1%    -4.81%  (p=0.008 n=5+5)
GetStripedOffsetsEven/length1000-8                                                                                      4.06ms ± 1%    4.05ms ± 0%      ~     (p=0.690 n=5+5)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5)-8                                                                  4.41ns ±13%    4.05ns ± 8%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5)-8                                                                       2.13µs ± 2%    2.10µs ± 3%      ~     (p=0.151 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5)_prefilled-8                                                             0.92ns ± 6%    0.92ns ± 7%      ~     (p=1.000 n=5+5)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5523)-8                                                               4.02ns ± 3%    4.19ns ±10%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5523)-8                                                                    2.24µs ± 4%    2.20µs ±15%      ~     (p=0.222 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5523)_prefilled-8                                                          0.90ns ± 9%    1.01ns ±19%      ~     (p=0.056 n=5+5)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5000000)-8                                                            4.06ns ± 3%    4.32ns ± 7%      ~     (p=0.151 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5000000)-8                                                                 2.13µs ± 2%    2.17µs ± 2%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5000000)_prefilled-8                                                       0.89ns ± 3%    1.00ns ± 6%   +11.41%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/segment.Scale(67280421310721)-8                                                     4.12ns ±11%    4.26ns ±10%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(67280421310721)-8                                                          2.12µs ± 2%    2.29µs ±24%      ~     (p=0.222 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(67280421310721)_prefilled-8                                                0.89ns ± 2%    1.02ns ± 5%   +14.65%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5)-8                                                               1.76µs ±11%    1.74µs ± 2%      ~     (p=0.286 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5)-8                                                                    2.38µs ± 5%    2.33µs ± 1%      ~     (p=0.286 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5)_prefilled-8                                                          0.89ns ± 4%    0.99ns ±20%   +12.16%  (p=0.032 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5523)-8                                                            1.69µs ± 1%    1.75µs ± 3%    +3.41%  (p=0.016 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5523)-8                                                                 2.34µs ± 2%    2.33µs ± 2%      ~     (p=0.690 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5523)_prefilled-8                                                       0.91ns ± 3%    1.00ns ±19%   +10.52%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5000000)-8                                                         1.75µs ±12%    1.74µs ± 3%      ~     (p=0.548 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5000000)-8                                                              2.38µs ±10%    2.33µs ± 2%      ~     (p=0.690 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5000000)_prefilled-8                                                    0.87ns ± 2%    0.91ns ± 6%      ~     (p=0.151 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(67280421310721)-8                                                  1.69µs ± 2%    1.70µs ± 3%      ~     (p=0.690 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(67280421310721)-8                                                       2.33µs ± 4%    2.27µs ± 1%      ~     (p=0.111 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(67280421310721)_prefilled-8                                             0.89ns ± 3%    0.92ns ± 2%      ~     (p=0.063 n=5+4)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/segment.Scale(5)-8                                  2.17µs ± 3%    2.20µs ± 3%      ~     (p=0.222 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5)-8                                       2.86µs ± 1%    2.82µs ± 1%    -1.11%  (p=0.048 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5)_prefilled-8                             9.90ns ± 2%    9.87ns ±11%      ~     (p=0.198 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/segment.Scale(5523)-8                               2.06µs ± 2%    2.10µs ± 3%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5523)-8                                    2.90µs ± 2%    2.87µs ± 2%      ~     (p=0.730 n=5+4)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5523)_prefilled-8                          8.71ns ± 4%    9.59ns ± 5%   +10.13%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/segment.Scale(5000000)-8                            1.81µs ± 1%    1.82µs ± 3%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5000000)-8                                 2.97µs ±11%    2.87µs ± 2%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5000000)_prefilled-8                       10.2ns ± 2%    10.7ns ± 6%      ~     (p=0.087 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/segment.Scale(67280421310721)-8                     1.96µs ± 4%    2.03µs ± 4%      ~     (p=0.095 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(67280421310721)-8                          2.88µs ± 2%    2.83µs ± 4%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(67280421310721)_prefilled-8                15.1ns ± 0%    15.3ns ± 2%    +1.76%  (p=0.016 n=4+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/segment.Scale(5)-8                      2.04µs ± 2%    2.04µs ± 2%      ~     (p=1.000 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5)-8                           3.48µs ± 1%    3.29µs ± 2%    -5.45%  (p=0.016 n=4+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5)_prefilled-8                 8.69ns ± 6%    8.32ns ± 9%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/segment.Scale(5523)-8                   1.96µs ± 2%    1.95µs ± 2%      ~     (p=0.310 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5523)-8                        3.49µs ± 4%    3.35µs ± 8%      ~     (p=0.095 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5523)_prefilled-8              9.20ns ±13%    8.31ns ± 4%      ~     (p=0.056 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/segment.Scale(5000000)-8                1.81µs ± 2%    1.80µs ± 1%      ~     (p=0.889 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5000000)-8                     3.44µs ± 2%    3.30µs ± 5%      ~     (p=0.056 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(5000000)_prefilled-8           10.1ns ± 2%    10.4ns ± 4%      ~     (p=0.111 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/segment.Scale(67280421310721)-8         1.97µs ± 4%    1.95µs ± 2%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(67280421310721)-8              3.48µs ± 2%    3.44µs ± 9%      ~     (p=0.690 n=5+5)
ExecutionSegmentScale/seq:0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.1/et.Scale(67280421310721)_prefilled-8    15.2ns ± 2%    15.2ns ± 2%      ~     (p=0.968 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/segment.Scale(5)-8                                                           2.64µs ± 9%    2.58µs ± 2%      ~     (p=0.841 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5)-8                                                                4.45µs ± 1%    4.46µs ± 2%      ~     (p=1.000 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5)_prefilled-8                                                      8.05ns ±12%    7.54ns ± 7%      ~     (p=0.151 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/segment.Scale(5523)-8                                                        2.75µs ± 1%    2.76µs ± 2%      ~     (p=0.738 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5523)-8                                                             4.53µs ± 7%    4.49µs ± 6%      ~     (p=0.548 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5523)_prefilled-8                                                   9.30ns ±11%    9.48ns ±13%      ~     (p=0.421 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/segment.Scale(5000000)-8                                                     2.56µs ± 3%    2.63µs ± 4%    +2.97%  (p=0.032 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5000000)-8                                                          4.48µs ± 3%    4.42µs ± 5%      ~     (p=0.222 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(5000000)_prefilled-8                                                10.4ns ± 0%    10.5ns ± 2%      ~     (p=0.635 n=4+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/segment.Scale(67280421310721)-8                                              2.88µs ± 2%    2.88µs ± 1%      ~     (p=1.000 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(67280421310721)-8                                                   4.47µs ± 3%    4.59µs ±11%      ~     (p=0.548 n=5+5)
ExecutionSegmentScale/seq:;segment:2/5:4/5/et.Scale(67280421310721)_prefilled-8                                         15.3ns ± 1%    15.1ns ± 2%      ~     (p=0.246 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/segment.Scale(5)-8                                                     2.68µs ± 2%    2.70µs ± 4%      ~     (p=0.952 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5)-8                                                           593µs ± 1%     562µs ± 1%    -5.26%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5)_prefilled-8                                                9.82ns ± 6%   26.90ns ± 4%  +173.87%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/segment.Scale(5523)-8                                                  2.93µs ± 6%    2.90µs ± 2%      ~     (p=0.794 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5523)-8                                                        594µs ± 1%     560µs ± 1%    -5.59%  (p=0.016 n=5+4)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5523)_prefilled-8                                             1.53µs ± 4%    0.03µs ± 3%   -98.33%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/segment.Scale(5000000)-8                                               2.82µs ± 9%    2.76µs ± 2%      ~     (p=0.841 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5000000)-8                                                     606µs ± 2%     563µs ± 1%    -7.11%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(5000000)_prefilled-8                                          5.94µs ± 2%    0.02µs ± 2%   -99.58%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/segment.Scale(67280421310721)-8                                        3.00µs ± 1%    2.95µs ± 3%      ~     (p=0.278 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(67280421310721)-8                                              599µs ± 1%     561µs ± 0%    -6.31%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:2235/5213:4/5/et.Scale(67280421310721)_prefilled-8                                   3.68µs ± 2%    0.03µs ± 5%   -99.27%  (p=0.008 n=5+5)

name                                                                                                                  old alloc/op   new alloc/op   delta
GetStripedOffsets/length10,seed777-8                                                                                    6.42kB ±12%    5.64kB ±10%      ~     (p=0.056 n=5+5)
GetStripedOffsets/length100,seed777-8                                                                                   49.8kB ± 7%    49.3kB ± 3%      ~     (p=0.690 n=5+5)
GetStripedOffsetsEven/length10-8                                                                                        1.49kB ± 0%    1.34kB ± 0%   -10.22%  (p=0.008 n=5+5)
GetStripedOffsetsEven/length100-8                                                                                       8.93kB ± 0%    7.24kB ± 0%   -18.91%  (p=0.008 n=5+5)
GetStripedOffsetsEven/length1000-8                                                                                      74.7kB ± 0%    58.5kB ± 0%   -21.67%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5)-8                                                                   0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/et.Scale(5)-8                                                                         624B ± 0%      608B ± 0%    -2.56%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5)_prefilled-8                                                              0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5523)-8                                                                0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/et.Scale(5523)-8                                                                      624B ± 0%      608B ± 0%    -2.56%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5523)_prefilled-8                                                           0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/segment.Scale(5000000)-8                                                             0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/et.Scale(5000000)-8                                                                   624B ± 0%      608B ± 0%    -2.56%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(5000000)_prefilled-8                                                        0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/segment.Scale(67280421310721)-8                                                      0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:/et.Scale(67280421310721)-8                                                            624B ± 0%      608B ± 0%    -2.56%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:/et.Scale(67280421310721)_prefilled-8                                                 0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5)-8                                                                 320B ± 0%      320B ± 0%      ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5)-8                                                                      720B ± 0%      704B ± 0%    -2.22%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5)_prefilled-8                                                           0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5523)-8                                                              320B ± 0%      320B ± 0%      ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5523)-8                                                                   720B ± 0%      704B ± 0%    -2.22%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5523)_prefilled-8                                                        0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(5000000)-8                                                           320B ± 0%      320B ± 0%      ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5000000)-8                                                                720B ± 0%      704B ± 0%    -2.22%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(5000000)_prefilled-8                                                     0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/segment.Scale(67280421310721)-8                                                    320B ± 0%      320B ± 0%      ~     (all equal)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(67280421310721)-8                                                         720B ± 0%      704B ± 0%    -2.22%  (p=0.008 n=5+5)
ExecutionSegmentScale/seq:;segment:0:1/et.Scale(67280421310721)_prefilled-8                                              0.00B          0.00B           ~     (all equal)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/segment.Scale(5)-8                                    424B ± 0%      424B ± 0%      ~     (all equal)
ExecutionSegmentScale/seq:0,0.3,0.5,0.6,0.7,0.8,0.9,1;segment:0:0.3/et.Scale(5)-8                                       1.06kB ± 0%    0.94kB ± 0%   -11.36%  (p=0.008 n=5+5)
…