@mstoykov 2022-11-15
I moved it to draft so people don't come looking it at it.

IMO it is still relevant, but arguably not useful in a lot of cases. It is mostly waiting for us to either hit problems with the current implementation or someone to have enough time to write better benchmarks and rerun them so we can be certain this actually improves stuff.

Both might happen the next time we refactor the executors or need to expand/implement a new one :shrug: 

But for now it just stays here